payloads are now correctly assigned to bins for the client. Tests successful

+++ psi_analytics_eurocrypt19/common/psi_analytics.h
@@ -34,7 +34,7 @@ namespace ENCRYPTO {
 uint64_t run_psi_analytics(const std::vector<std::uint64_t> &inputs, PsiAnalyticsContext &context);
 
 uint64_t run_psi_analytics(const std::vector<std::uint64_t> &inputs, PsiAnalyticsContext &context,
-                           const std::vector<std::uint8_t> &payload_input_a);
+                           const std::vector<std::uint32_t> &payload_input_a);
 
 std::vector<std::pair<uint64_t, uint64_t>> OpprgPsiClient(const std::vector<uint64_t> &elements,
                                                           PsiAnalyticsContext &context);
